@@ -34,24 +34,24 @@
     "rc-menu": "^5.0.10",
     "rc-progress": "^2.1.2",
     "rc-switch": "^1.4.2",
-    "react": "^15.4.2",
+    "react": "^16.2",
     "react-autolink-text2": "git+https://github.com/xodio/react-autolink-text.git#npm",
     "react-autosuggest": "git+https://github.com/xodio/react-autosuggest.git#bc6ab71a2ec47ad5fbce5d13ff6a572942ea765e",
     "react-click-outside": "^2.2.0",
     "react-codemirror": "^1.0.0",
-    "react-collapsible": "git+https://github.com/xodio/react-collapsible.git#npm",
+    "react-collapsible": "^2.0.3",
     "react-custom-scroll": "^2.0.0",
     "react-dnd": "^2.5.1",
     "react-dnd-html5-backend": "^2.5.1",
-    "react-dom": "^15.4.2",
+    "react-dom": "^16.2",
     "react-event-listener": "^0.4.5",
     "react-fa": "^5.0.0",
     "react-highlight-words": "^0.8.1",
-    "react-hotkeys": "git+https://github.com/xodio/react-hotkeys.git#npm",
+    "react-hotkeys": "^0.10.0",
     "react-infinite": "^0.12.1",
     "react-redux": "^4.0.6",
     "react-skylight": "git+https://github.com/xodio/react-skylight.git",
-    "react-sortable-hoc": "0.0.7",
+    "react-sortable-hoc": "0.6.7",
     "react-split-pane": "^0.1.68",
     "recompose": "^0.25.0",
     "redux": "^3.0.5",
